[remove] utils 폴더 생성 및 getRandomCards 이동
wq
